Reduce the size of the JSON files (whitespace-only changes) (#135)
* Remove redundant whitespace at EOL from the JSON files.

This saves over 12½ KB in these files; keeping the sizes small
helps applications that embed those files “as-is” into their
binaries (as opposed to, in some parsed form), e.g. MuseScore.

* Correct double space in the description of two glyphs.

* Indent the JSON files with tabs.

See commit f4f7a15955583e658001612c3ced267d56b96848 for the
rationale of this change. This saves over 103 KB; with this
change series, the files are about 20% smaller.